DFR-1338 added tests
