+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/ConsentOrderApprovedControllerTest.java
@@ -1,331 +1,87 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.controllers;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.Before;
 import org.junit.Test;
-import org.mockito.stubbing.OngoingStubbing;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.ResultActions;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.config.DocumentConfiguration;
+import uk.gov.hmcts.reform.finrem.caseorchestration.helper.ConsentedApplicationHelper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.PensionTypeCollection;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper.ConsentOrderWrapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.CaseDataService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.ConsentOrderApprovedDocumentService;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.ConsentOrderPrintService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.GenericDocumentService;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.NotificationService;
 
-import java.util.List;
-
-import static java.util.Collections.singletonList;
-import static org.hamcrest.Matchers.hasKey;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.not;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.isA;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.AUTHORIZATION_HEADER;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TOKEN;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.BINARY_URL;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.DOC_URL;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.FILE_NAME;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.PENSION_TYPE;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.caseDocument;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.feignError;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.pensionDocumentData;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper.DOCUMENT_BINARY_URL;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ConsentedStatus.CONSENT_ORDER_MADE;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.LATEST_CONSENT_ORDER;
 
-@WebMvcTest(ConsentOrderApprovedController.class)
-public class ConsentOrderApprovedControllerTest extends BaseControllerTest {
+@RunWith(MockitoJUnitRunner.class)
+public class ConsentOrderApprovedControllerTest {
 
-    @MockBean
-    private ConsentOrderApprovedDocumentService consentOrderApprovedDocumentService;
-    @MockBean
+    @Mock
     private GenericDocumentService genericDocumentService;
-    @MockBean
-    private ConsentOrderPrintService consentOrderPrintService;
-    @MockBean
-    private NotificationService notificationService;
-    @MockBean
+    @Mock
+    private DocumentConfiguration documentConfiguration;
+    @Mock
     private DocumentHelper documentHelper;
-    @MockBean
+
+    private ObjectMapper mapper = new ObjectMapper();
+    @Mock
     private CaseDataService caseDataService;
+    @Mock
+    private ConsentedApplicationHelper consentedApplicationHelper;
+    @InjectMocks
+    private ConsentOrderApprovedDocumentService consentOrderApprovedDocumentService;
 
-    public String consentOrderApprovedEndpoint() {
-        return "/case-orchestration/documents/consent-order-approved";
+
+    @Before
+    public void setup() {
+        consentOrderApprovedDocumentService = new ConsentOrderApprovedDocumentService(genericDocumentService,
+            documentConfiguration, documentHelper, mapper, caseDataService, consentedApplicationHelper);
     }
 
-    public String contestedConsentOrderApprovedEndpoint() {
-        return "/case-orchestration/consent-in-contested/consent-order-approved";
-    }
-
-    public String contestedConsentSendOrderEndpoint() {
-        return "/case-orchestration/consent-in-contested/send-order";
-    }
 
     @Test
-    public void consentOrderApproved400Error() throws Exception {
-        doEmptyCaseDataSetUp();
+    public void givenFinremCaseDetails_whenAddGeneratedApprConsOrderDocsToCase_thenGenerateAndAddDocsToCase() {
+        CaseDocument uploadApproveOrder = CaseDocument.builder().documentFilename("testUploadAppOrder").build();
+        FinremCaseData finremCaseData = FinremCaseData.builder().consentOrderWrapper(
+                ConsentOrderWrapper.builder().uploadApprovedConsentOrder(uploadApproveOrder).build())
+            .build();
 
-        mvc.perform(post(consentOrderApprovedEndpoint())
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isBadRequest());
+        when(genericDocumentService.annexStampDocument(any(), any()))
+            .thenReturn(uploadApproveOrder);
+        when(genericDocumentService.generateDocument(any(), any(), any(), any()))
+            .thenReturn(uploadApproveOrder);
+        when(documentHelper.deepCopy(any(), any()))
+            .thenReturn(toCaseDetails(FinremCaseDetails.builder().data(finremCaseData).build()));
+
+        FinremCaseDetails finremCaseDetails = FinremCaseDetails.builder().id(1L).data(finremCaseData).build();
+        consentOrderApprovedDocumentService.addGeneratedApprovedConsentOrderDocumentsToCase("AUTH_TOKEN",
+            finremCaseDetails);
+
+        assertThat(finremCaseDetails.getData().getApprovedOrderCollection().get(0).getApprovedOrder()
+                .getConsentOrder().getDocumentFilename(),
+            is("testUploadAppOrder"));
     }
 
-    @Test
-    public void consentOrderApproved500Error() throws Exception {
-        doValidCaseDataSetUp();
-        whenServiceGeneratesDocument().thenThrow(feignError());
+    private CaseDetails toCaseDetails(FinremCaseDetails finremCaseDetails) {
+        CaseDetails generateDocumentPayload = null;
+        try {
+            generateDocumentPayload = mapper.readValue(mapper.writeValueAsString(finremCaseDetails), CaseDetails.class);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
 
-        mvc.perform(post(consentOrderApprovedEndpoint())
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isInternalServerError());
-    }
-
-    @Test
-    public void latestConsentOrderIsMissing() throws Exception {
-        doMissingLatestConsentOrder();
-        whenServiceGeneratesDocument().thenReturn(caseDocument());
-        whenServiceGeneratesNotificationLetter().thenReturn(caseDocument());
-        whenAnnexStampingDocument().thenReturn(caseDocument());
-        whenStampingDocument().thenReturn(caseDocument());
-        whenStampingPensionDocuments().thenReturn(singletonList(pensionDocumentData()));
-
-        ResultActions result = mvc.perform(post(consentOrderApprovedEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk());
-        result.andExpect(jsonPath("$.data", not(hasKey(LATEST_CONSENT_ORDER))));
-    }
-
-    @Test
-    public void consentOrderApprovedSuccess() throws Exception {
-        doValidCaseDataSetUp();
-        whenServiceGeneratesDocument().thenReturn(caseDocument());
-        whenServiceGeneratesNotificationLetter().thenReturn(caseDocument());
-        whenAnnexStampingDocument().thenReturn(caseDocument());
-        whenStampingDocument().thenReturn(caseDocument());
-        whenStampingPensionDocuments().thenReturn(singletonList(pensionDocumentData()));
-        when(documentHelper.getPensionDocumentsData(any())).thenReturn(singletonList(caseDocument()));
-
-        ResultActions result = mvc.perform(post(consentOrderApprovedEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk());
-        assertLetter(result);
-        assertConsentOrder(result);
-        assertPensionDocs(result);
-    }
-
-    @Test
-    public void consentOrderApprovedSuccessForPaperApplication() throws Exception {
-        doValidCaseDataSetUpForPaperApplication();
-        whenServiceGeneratesDocument().thenReturn(caseDocument());
-        whenServiceGeneratesNotificationLetter().thenReturn(caseDocument());
-        whenAnnexStampingDocument().thenReturn(caseDocument());
-        whenStampingDocument().thenReturn(caseDocument());
-        whenStampingPensionDocuments().thenReturn(singletonList(pensionDocumentData()));
-        when(documentHelper.getPensionDocumentsData(any())).thenReturn(singletonList(caseDocument()));
-
-        ResultActions result = mvc.perform(post(consentOrderApprovedEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk());
-        assertLetter(result);
-        assertConsentOrder(result);
-        assertPensionDocs(result);
-    }
-
-    @Test
-    public void shouldUpdateStateToConsentOrderMadeAndBulkPrint() throws Exception {
-        doValidCaseDataSetUpNoPensionCollection();
-        whenServiceGeneratesDocument().thenReturn(caseDocument());
-        whenServiceGeneratesNotificationLetter().thenReturn(caseDocument());
-        whenAnnexStampingDocument().thenReturn(caseDocument());
-        whenStampingDocument().thenReturn(caseDocument());
-        whenStampingPensionDocuments().thenReturn(singletonList(pensionDocumentData()));
-        when(notificationService.isRespondentSolicitorEmailCommunicationEnabled(any())).thenReturn(true);
-        when(caseDataService.isApplicantSolicitorAgreeToReceiveEmails(any())).thenReturn(true);
-
-        ResultActions result = mvc.perform(post(consentOrderApprovedEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk())
-            .andExpect(jsonPath("$.data.state", is(CONSENT_ORDER_MADE.toString())));
-
-        verify(consentOrderPrintService).sendConsentOrderToBulkPrint(any(), any());
-        verify(notificationService).sendConsentOrderAvailableCtscEmail(any());
-        verify(notificationService).sendConsentOrderAvailableEmailToApplicantSolicitor(any());
-        verify(notificationService).sendConsentOrderAvailableEmailToRespondentSolicitor(any());
-    }
-
-    @Test
-    public void shouldUpdateStateToConsentOrderMadeAndBulkPrint_noEmails() throws Exception {
-        doValidCaseDataSetUpNoPensionCollection();
-        whenServiceGeneratesDocument().thenReturn(caseDocument());
-        whenServiceGeneratesNotificationLetter().thenReturn(caseDocument());
-        whenAnnexStampingDocument().thenReturn(caseDocument());
-        whenStampingDocument().thenReturn(caseDocument());
-        whenStampingPensionDocuments().thenReturn(singletonList(pensionDocumentData()));
-        when(notificationService.isRespondentSolicitorEmailCommunicationEnabled(any())).thenReturn(false);
-        when(caseDataService.isApplicantSolicitorAgreeToReceiveEmails(any())).thenReturn(false);
-
-        ResultActions result = mvc.perform(post(consentOrderApprovedEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk())
-            .andExpect(jsonPath("$.data.state", is(CONSENT_ORDER_MADE.toString())));
-
-        verify(consentOrderPrintService).sendConsentOrderToBulkPrint(any(), any());
-        verify(notificationService).sendConsentOrderAvailableCtscEmail(any());
-        verify(notificationService, never()).sendConsentOrderAvailableEmailToApplicantSolicitor(any());
-        verify(notificationService, never()).sendConsentOrderAvailableEmailToRespondentSolicitor(any());
-    }
-
-    @Test
-    public void shouldNotTriggerConsentOrderApprovedNotificationLetterIfIsNotPaperApplication() throws Exception {
-        doValidCaseDataSetUp();
-
-        whenServiceGeneratesDocument().thenReturn(caseDocument());
-        whenServiceGeneratesNotificationLetter().thenReturn(caseDocument());
-        whenAnnexStampingDocument().thenReturn(caseDocument());
-        whenStampingDocument().thenReturn(caseDocument());
-        whenStampingPensionDocuments().thenReturn(singletonList(pensionDocumentData()));
-        when(documentHelper.getPensionDocumentsData(any())).thenReturn(singletonList(caseDocument()));
-
-        ResultActions result = mvc.perform(post(consentOrderApprovedEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk());
-        assertLetter(result);
-        assertConsentOrder(result);
-        assertPensionDocs(result);
-        verify(consentOrderApprovedDocumentService, never()).generateApprovedConsentOrderCoverLetter(any(), any());
-    }
-
-    @Test
-    public void consentInContestedConsentOrderApprovedShouldProcessDocuments() throws Exception {
-        doValidCaseDataSetUp();
-
-        mvc.perform(post(contestedConsentOrderApprovedEndpoint())
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk());
-
-        verify(consentOrderApprovedDocumentService, times(1)).stampAndPopulateContestedConsentApprovedOrderCollection(any(), eq(AUTH_TOKEN));
-        verify(consentOrderApprovedDocumentService, times(1)).generateAndPopulateConsentOrderLetter(any(), eq(AUTH_TOKEN));
-    }
-
-    @Test
-    public void consentInContestedConsentOrderApprovedShouldProcessPensionDocs() throws Exception {
-        doValidConsentInContestWithPensionData();
-
-        mvc.perform(post(contestedConsentOrderApprovedEndpoint())
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk());
-
-        verify(consentOrderApprovedDocumentService).stampAndPopulateContestedConsentApprovedOrderCollection(any(), eq(AUTH_TOKEN));
-    }
-
-    @Test
-    public void consentInContestedSendOrderShouldPrintDocsWhenNotApproved() throws Exception {
-        doValidCaseDataSetUp();
-
-        mvc.perform(post(contestedConsentSendOrderEndpoint())
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk());
-
-        verify(consentOrderApprovedDocumentService, never()).generateApprovedConsentOrderLetter(any(), eq(AUTH_TOKEN));
-        verify(consentOrderPrintService).sendConsentOrderToBulkPrint(any(), eq(AUTH_TOKEN));
-    }
-
-    @Test
-    public void consentInContestedSendOrderShouldPrintDocsWhenApproved() throws Exception {
-        doValidConsentOrderApprovedSetup();
-        when(consentOrderApprovedDocumentService.generateApprovedConsentOrderLetter(any(), eq(AUTH_TOKEN)))
-            .thenReturn(caseDocument());
-
-        ResultActions result = mvc.perform(post(contestedConsentSendOrderEndpoint())
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        result.andExpect(status().isOk());
-        verify(consentOrderPrintService).sendConsentOrderToBulkPrint(any(), eq(AUTH_TOKEN));
-    }
-
-    private OngoingStubbing<CaseDocument> whenServiceGeneratesDocument() {
-        return when(consentOrderApprovedDocumentService.generateApprovedConsentOrderLetter(isA(CaseDetails.class), eq(AUTH_TOKEN)));
-    }
-
-    private OngoingStubbing<CaseDocument> whenServiceGeneratesNotificationLetter() {
-        return when(consentOrderApprovedDocumentService.generateApprovedConsentOrderCoverLetter(isA(CaseDetails.class), eq(AUTH_TOKEN)));
-    }
-
-    private OngoingStubbing<CaseDocument> whenAnnexStampingDocument() {
-        return when(genericDocumentService.annexStampDocument(isA(CaseDocument.class), eq(AUTH_TOKEN)));
-    }
-
-    private OngoingStubbing<CaseDocument> whenStampingDocument() {
-        return when(genericDocumentService.stampDocument(isA(CaseDocument.class), eq(AUTH_TOKEN)));
-    }
-
-    private OngoingStubbing<List<PensionTypeCollection>> whenStampingPensionDocuments() {
-        return when(consentOrderApprovedDocumentService.stampPensionDocuments(any(), eq(AUTH_TOKEN)));
-    }
-
-    private void assertLetter(ResultActions result) throws Exception {
-        String path = "$.data.approvedOrderCollection[0].value.orderLetter.";
-        result.andExpect(jsonPath(path + "document_url", is(DOC_URL)))
-            .andExpect(jsonPath(path + "document_filename", is(FILE_NAME)))
-            .andExpect(jsonPath(path + DOCUMENT_BINARY_URL, is(BINARY_URL)));
-    }
-
-    private void assertConsentOrder(ResultActions result) throws Exception {
-        assertDocument(result, "$.data.approvedOrderCollection[0].value.consentOrder.");
-    }
-
-    private void assertPensionDocs(ResultActions result) throws Exception {
-        String path = "$.data.approvedOrderCollection[0].value.pensionDocuments[0].value.";
-        String docPath = "$.data.approvedOrderCollection[0].value.pensionDocuments[0].value.uploadedDocument.";
-        result.andExpect(jsonPath(path + "typeOfDocument", is(PENSION_TYPE)));
-        assertDocument(result, docPath);
-    }
-
-    private void assertDocument(ResultActions result, String path) throws Exception {
-        result.andExpect(jsonPath(path + "document_url", is(DOC_URL)))
-            .andExpect(jsonPath(path + "document_filename", is(FILE_NAME)))
-            .andExpect(jsonPath(path + DOCUMENT_BINARY_URL, is(BINARY_URL)));
+        return generateDocumentPayload;
     }
 }
